Renamed prototype Shared namespace to Players
