+++ Assets/Prototype/LiteNetLib/Players/Packets.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e6b957c6a348dac49870cdc62ec19b87
+guid: ede1befb8482f4341814a8b2e961c32c
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
