+++ Assets/Prototype/LiteNetLib/Players/Packets/PlayerCreatePacket.cs
@@ -3,7 +3,7 @@ using Exanite.Arpg.Networking;
 using LiteNetLib.Utils;
 using UnityEngine;
 
-namespace Prototype.LiteNetLib.Shared.Packets
+namespace Prototype.LiteNetLib.Players.Packets
 {
     public class PlayerCreatePacket : IPacket
     {
