+++ Assets/Prototype/LiteNetLib/Players/Packets/PlayerCreatePacket.cs.meta
