+++ Assets/Prototype/LiteNetLib/Players/Packets/PlayerDestroyPacket.cs
@@ -1,9 +1,9 @@
 ﻿using Exanite.Arpg.Networking;
 using LiteNetLib.Utils;
 
-namespace Prototype.LiteNetLib.Shared.Packets
+namespace Prototype.LiteNetLib.Players.Packets
 {
-    public class PlayerIdAssignmentPacket : IPacket
+    public class PlayerDestroyPacket : IPacket
     {
         public int id;
