+++ Assets/Prototype/LiteNetLib/Players/Player.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7cfd211044309184fba2d2923f1fa6b7
+guid: c08eba04a0090c942a66d385c907c012
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
