+++ Assets/Prototype/LiteNetLib/Players/PlayerConnection.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0650de482360cc044a93aba396510cab
+guid: 14b0ece8857f06e469341cfcbfcba2d1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
