+++ Assets/Prototype/LiteNetLib/Players/PlayerManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2553d0259a233bc45b601082d47c6afa
+guid: 6527654db4ab0c34f87b8b09bfc4d898
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
